Add more rat bc tests

+++ r600/cf_node.h
@@ -239,7 +239,7 @@ public:
                   uint16_t rw_gpr,
                   uint16_t index_gpr,
                   uint16_t elem_size,
-                  uint16_t array_size,
+                  uint32_t array_size,
                   uint16_t comp_mask,
                   uint16_t burst_count,
                   const cf_flags &flags);
@@ -250,7 +250,7 @@ protected:
    virtual void print_export_detail(std::ostream& os) const = 0;
    virtual void encode_export_parts(uint64_t& bc) const = 0;
 
-   uint16_t m_array_size;
+   uint32_t m_array_size;
    uint16_t m_comp_mask;
 };
 
@@ -265,7 +265,7 @@ public:
                uint16_t rw_gpr,
                uint16_t index_gpr,
                uint16_t elem_size,
-               uint16_t array_size,
+               uint32_t array_size,
                uint16_t comp_mask,
                uint16_t burst_count,
                const cf_flags &flags);
